카테고리 목록 fetch하는 것을 컴포넌트 분리하여 Suspense와 에러 바운더리를 적용 (#352)
* refactor: (#324) 대쉬보드에서 카테고리를 불러오고 밖에서 suspense를 이용하도록 수정
대쉬보드 스토리북 사용하지 않는 코드 삭제

* feat: (#324) 에러 바운더리에서 에러 메세지를 렌더하도록 수정
대쉬보드에서 카테고리 불러오는 것을 실패할 때 에러바운더리에서 에러를 잡도록 구현

* fix: (#324) MSW에서 카테고리 리스트를 불러올 때 딜레이가 너무 길어 실패하는 것을 수정

* refactor: (#324) 현재 선택된 게시글 종류 보여주는 로직 PostList,Layout에서 Dashboard로 이동

* refactor: (#324) CategorySection 컴포넌트에서 카테고리 목록을 불러와 보여주도록 구현

* refactor: (#324) ??로 카테고리 리스트가 없다면 빈 배열로 설정하는 중복되는 코드를 제거

* refactor: (#324) 사용하지 않는 코드 삭제 및 스타일드 컴포넌트명 변경
