+++ frontend/src/components/common/Drawer/Drawer.stories.tsx
@@ -1,8 +1,5 @@
 import type { Meta } from '@storybook/react';
 
-import { Category } from '@type/category';
-import { User } from '@type/user';
-
 import { useDrawer } from '@hooks/useDrawer';
 
 import Dashboard from '../Dashboard';
@@ -16,21 +13,6 @@ const meta: Meta<typeof Drawer> = {
 
 export default meta;
 
-const MOCK_USER_INFO: User = {
-  nickname: '우아한 코끼리',
-  postCount: 4,
-  voteCount: 128,
-  gender: 'MALE',
-  birthYear: 1997,
-};
-
-const MOCK_CATEGORIES: Category[] = [
-  { id: 12312, name: '음식', isFavorite: false },
-  { id: 12, name: '연애', isFavorite: false },
-  { id: 13, name: '패션', isFavorite: false },
-  { id: 14, name: '금융', isFavorite: false },
-];
-
 export const LeftSideBar = () => {
   const { drawerRef, openDrawer, closeDrawer } = useDrawer('left');
 
@@ -38,12 +20,7 @@ export const LeftSideBar = () => {
     <div>
       <NarrowMainHeader handleMenuOpenClick={openDrawer} />
       <Drawer width="225px" handleDrawerClose={closeDrawer} placement="left" ref={drawerRef}>
-        <Dashboard
-          userInfo={MOCK_USER_INFO}
-          categoryList={MOCK_CATEGORIES}
-          handleLogoutClick={() => {}}
-          selectedState="전체"
-        />
+        <Dashboard />
       </Drawer>
     </div>
   );
@@ -56,12 +33,7 @@ export const RightSideBar = () => {
     <div>
       <NarrowMainHeader handleMenuOpenClick={openDrawer} />
       <Drawer width="225px" handleDrawerClose={closeDrawer} placement="right" ref={drawerRef}>
-        <Dashboard
-          userInfo={MOCK_USER_INFO}
-          categoryList={MOCK_CATEGORIES}
-          handleLogoutClick={() => {}}
-          selectedState="전체"
-        />
+        <Dashboard />
       </Drawer>
     </div>
   );
