+++ frontend/src/components/common/Layout/index.tsx
@@ -1,15 +1,8 @@
-import { PropsWithChildren, useContext } from 'react';
+import { PropsWithChildren } from 'react';
 import { useNavigate } from 'react-router-dom';
-
-import { AuthContext } from '@hooks/context/auth';
-import { useCategoryList } from '@hooks/query/category/useCategoryList';
-import { usePostRequestInfo } from '@hooks/usePostRequestInfo';
 
 import Dashboard from '@components/common/Dashboard';
 import WideHeader from '@components/common/WideHeader';
-
-import { clearCookieToken } from '@utils/cookie';
-import { getSelectedState } from '@utils/post/getSelectedState';
 
 import * as S from './style';
 
@@ -19,24 +12,6 @@ interface LayoutProps extends PropsWithChildren {
 
 export default function Layout({ children, isSidebarVisible }: LayoutProps) {
   const navigate = useNavigate();
-
-  const { loggedInfo, clearLoggedInfo } = useContext(AuthContext);
-
-  const { data: categoryList } = useCategoryList(loggedInfo.isLoggedIn);
-  const { postOptionalOption, postType } = usePostRequestInfo();
-  const { categoryId, keyword } = postOptionalOption;
-
-  const selectedState = getSelectedState({
-    categoryId,
-    keyword,
-    categoryList: categoryList ?? [],
-    postType,
-  });
-
-  const handleLogoutClick = () => {
-    clearCookieToken('accessToken');
-    clearLoggedInfo();
-  };
 
   const movePostListPage = () => {
     navigate('/');
@@ -50,12 +25,7 @@ export default function Layout({ children, isSidebarVisible }: LayoutProps) {
       <S.ContentContainer>
         {isSidebarVisible && (
           <S.DashboardWrapper>
-            <Dashboard
-              userInfo={loggedInfo.userInfo}
-              categoryList={categoryList ?? []}
-              selectedState={selectedState}
-              handleLogoutClick={handleLogoutClick}
-            />
+            <Dashboard />
           </S.DashboardWrapper>
         )}
         <S.MainContainer $isSidebarVisible={isSidebarVisible}>
