+++ frontend/src/components/post/PostListPage/index.tsx
@@ -1,9 +1,6 @@
-import { Suspense, useContext } from 'react';
+import { Suspense } from 'react';
 
-import { AuthContext } from '@hooks/context/auth';
-import { useCategoryList } from '@hooks/query/category/useCategoryList';
 import { useDrawer } from '@hooks/useDrawer';
-import { usePostRequestInfo } from '@hooks/usePostRequestInfo';
 
 import ErrorBoundary from '@pages/ErrorBoundary';
 
@@ -17,27 +14,12 @@ import PostList from '@components/post/PostList';
 
 import { PATH } from '@constants/path';
 
-import { getSelectedState } from '@utils/post/getSelectedState';
 import { scrollToTop } from '@utils/scrollToTop';
 
 import * as S from './style';
 
 export default function PostListPage() {
   const { drawerRef, closeDrawer, openDrawer } = useDrawer('left');
-
-  const { isLoggedIn: isLogged, userInfo } = useContext(AuthContext).loggedInfo;
-  const { data: categoryList } = useCategoryList(isLogged);
-  const { postOptionalOption, postType } = usePostRequestInfo();
-  const { categoryId, keyword } = postOptionalOption;
-
-  const selectedState = getSelectedState({
-    categoryId,
-    keyword,
-    categoryList: categoryList ?? [],
-    postType,
-  });
-
-  const handleLogoutClick = () => {};
 
   return (
     <S.Container>
@@ -46,12 +28,7 @@ export default function PostListPage() {
       </S.HeaderWrapper>
       <S.DrawerWrapper>
         <Drawer handleDrawerClose={closeDrawer} placement="left" width="225px" ref={drawerRef}>
-          <Dashboard
-            userInfo={userInfo}
-            categoryList={categoryList ?? []}
-            handleLogoutClick={handleLogoutClick}
-            selectedState={selectedState}
-          />
+          <Dashboard />
         </Drawer>
       </S.DrawerWrapper>
       <ErrorBoundary>
